Inline marked environment detection
The checks for `global` and `define` and such are incompatible with the Google Closure Compiler running in Advanced mode. The checks also weren't used; in practice, it would always set `marked` to be `module.exports` anyway.

This inlines the `module.exports` logic such that it returns `marked` directly rather than doing the `exposing` dance. Besides being slightly more efficient, this also makes it so that the Google Closure Compiler can be run on Elm projects that depend on this package in Advanced mode.

This also adds a comment to the hljs section explaining how to get it to work with Closure Compiler running in Advanced Mode.